requirements.txt
Django==1.6
dj-static==0.0.5
gunicorn==18.0
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2
djangorestframework==2.3.14
django-cors-headers==0.12